refactor: extract submissions dashboard to own component
- update main view naming conventions for clarity
- #120 #125
